feat(client): simplify client; removes high-level & cache
BREAKING CHANGE: exports single `Client` (removes LowLevel and HighLevelClient) and removes Cache functionality.
