+++ src/index.ts
@@ -1,3 +1,1 @@
-export { Cache } from './cache'
 export { Client } from './client'
-export { ClientLowLevel } from './clientLowLevel'
