+++ test/authlessClient.spec.ts
@@ -1,20 +1,21 @@
 /* eslint-env node, jest */
-import { Client, ClientLowLevel } from '../src'
+import { Client } from '../src'
 
 describe('Client', () => {
   let authlessClient: Client | null = null
-  let authlessLowLevelClient: ClientLowLevel | null = null
-  beforeEach(() => {
-    authlessLowLevelClient = new ClientLowLevel({
-      retries: 0,
-      uri: 'http://example.com:4000',
-    })
-  })
 
   describe('new', () => {
+    test('rejects incorrect constructor config', () => {
+      const invalidConfig = {
+        serverUri: 'http://example.com:4000',
+        retries: 's' as unknown as number,
+      }
+      expect(() => new Client(invalidConfig)).toThrow()
+    })
+
     test('returns authlessClient instance', () => {
       authlessClient = new Client({
-        lowLevelClient: authlessLowLevelClient,
+        serverUri: 'http://example.com:4000',
         retries: 1,
       })
       expect(authlessClient).toBeInstanceOf(Client)
